Forth commit
Updated the app icon to fit within the borders.
Updated first and last name textfields to capitalize the first letter. Added an image to meet the image requirement of the assignment. First image App Icon, Second image located in add a contact activity, Third image located in the contacts list activity to the left of each added contact.